+++ idp_engine/Theory.py
@@ -582,6 +582,8 @@ class Theory(object):
                 count += 1
                 ass = self._from_model(solver, todo, complete)
                 yield ass
+                if not todo:  # empty theory
+                    break
             else:
                 break